refactor(utils): move ghdl_msg parsing to its own module
The change should help separate easier to test
ghdl message parsing and building from the
actual test runner and renderer, that rely
heavily on the interaction with ghdl
and checking the output.
